# split out null value index