--- a/processing/src/main/java/org/apache/druid/query/dimension/PredicateFilteredDimensionSelector.java
+++ b/processing/src/main/java/org/apache/druid/query/dimension/PredicateFilteredDimensionSelector.java
@@ -64,7 +64,6 @@ final class PredicateFilteredDimensionSelector extends AbstractDimensionSelector
   @Override
   public ValueMatcher makeValueMatcher(final String value)
   {
-    final boolean matchNull = predicate.apply(null);
     return new ValueMatcher()
     {
       @Override
@@ -82,8 +81,8 @@ final class PredicateFilteredDimensionSelector extends AbstractDimensionSelector
             nullRow = false;
           }
         }
-        // null should match empty rows in multi-value columns if predicate matches null
-        return nullRow && value == null && matchNull;
+        // null should match empty rows in multi-value columns
+        return nullRow && value == null;
       }
 
       @Override
@@ -98,7 +97,7 @@ final class PredicateFilteredDimensionSelector extends AbstractDimensionSelector
   @Override
   public ValueMatcher makeValueMatcher(final Predicate<String> matcherPredicate)
   {
-    final boolean matchNull = predicate.apply(null) && matcherPredicate.apply(null);
+    final boolean matchNull = matcherPredicate.apply(null);
     return new ValueMatcher()
     {
       @Override