--- a/processing/src/main/java/org/apache/druid/segment/column/ColumnBuilder.java
+++ b/processing/src/main/java/org/apache/druid/segment/column/ColumnBuilder.java
@@ -21,8 +21,11 @@ package org.apache.druid.segment.column;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Supplier;
+import org.apache.druid.collections.bitmap.ImmutableBitmap;
+import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.io.smoosh.SmooshedFileMapper;
 import org.apache.druid.segment.serde.NoIndexesColumnIndexSupplier;
+import org.apache.druid.segment.serde.NullValueIndexSupplier;
 
 import javax.annotation.Nullable;
 
@@ -71,6 +74,7 @@ public class ColumnBuilder
 
   public ColumnBuilder setDictionaryEncodedColumnSupplier(Supplier<? extends DictionaryEncodedColumn<?>> columnSupplier)
   {
+    checkColumnSupplierNotSet();
     this.columnSupplier = columnSupplier;
     this.capabilitiesBuilder.setDictionaryEncoded(true);
     this.capabilitiesBuilder.setDictionaryValuesSorted(true);
@@ -87,12 +91,14 @@ public class ColumnBuilder
 
   public ColumnBuilder setComplexColumnSupplier(Supplier<? extends ComplexColumn> columnSupplier)
   {
+    checkColumnSupplierNotSet();
     this.columnSupplier = columnSupplier;
     return this;
   }
 
   public ColumnBuilder setNumericColumnSupplier(Supplier<? extends NumericColumn> columnSupplier)
   {
+    checkColumnSupplierNotSet();
     this.columnSupplier = columnSupplier;
     return this;
   }
@@ -103,9 +109,17 @@ public class ColumnBuilder
       boolean hasSpatial
   )
   {
+    checkIndexSupplierNotSet();
     this.indexSupplier = indexSupplier;
     capabilitiesBuilder.setHasBitmapIndexes(hasBitmapIndex);
     capabilitiesBuilder.setHasSpatialIndexes(hasSpatial);
+    return this;
+  }
+
+  public ColumnBuilder setNullValueIndexSupplier(ImmutableBitmap nullValueIndex)
+  {
+    checkIndexSupplierNotSet();
+    this.indexSupplier = new NullValueIndexSupplier(nullValueIndex);
     return this;
   }
 
@@ -125,5 +139,20 @@ public class ColumnBuilder
     Preconditions.checkState(capabilitiesBuilder.getType() != null, "Type must be set.");
 
     return new SimpleColumnHolder(capabilitiesBuilder, columnSupplier, indexSupplier);
+  }
+
+  private void checkColumnSupplierNotSet()
+  {
+    if (columnSupplier != null) {
+      throw new ISE("Column supplier already set!");
+    }
+  }
+
+  private void checkIndexSupplierNotSet()
+  {
+    //noinspection ObjectEquality
+    if (indexSupplier != NoIndexesColumnIndexSupplier.getInstance()) {
+      throw new ISE("Index supplier already set!");
+    }
   }
 }