--- a/processing/src/main/java/org/apache/druid/segment/column/LexicographicalRangeIndex.java
+++ b/processing/src/main/java/org/apache/druid/segment/column/LexicographicalRangeIndex.java
@@ -25,12 +25,15 @@ import javax.annotation.Nullable;
 
 /**
  * An optimized column value {@link BitmapColumnIndex} provider for columns which are stored in 'lexicographical' order,
- * allowing short-circuit processing of string value ranges.
+ * allowing short-circuit processing of string value ranges. This index does not match null values, union the results
+ * of this index with {@link NullValueIndex} if null values should be considered part of the value range.
  */
 public interface LexicographicalRangeIndex
 {
   /**
-   * Get a {@link BitmapColumnIndex} corresponding to the values supplied in the specified range.
+   * Get a {@link BitmapColumnIndex} corresponding to the values supplied in the specified range. If supplied starting
+   * value is null, the range will begin at the first non-null value in the underlying value dictionary. If the end
+   * value is null, the range will extend to the last value in the underlying value dictionary.
    */
   BitmapColumnIndex forRange(
       @Nullable String startValue,
@@ -41,7 +44,9 @@ public interface LexicographicalRangeIndex
 
   /**
    * Get a {@link BitmapColumnIndex} corresponding to the values supplied in the specified range whose dictionary ids
-   * also match some predicate, such as to match a prefix.
+   * also match some predicate, such as to match a prefix. If supplied starting value is null, the range will begin at
+   * the first non-null value in the underlying value dictionary that matches the predicate. If the end value is null,
+   * the range will extend to the last value in the underlying value dictionary that matches the predicate.
    *
    * If the provided {@code} matcher is always true, it's better to use the other
    * {@link #forRange(String, boolean, String, boolean)} method.