--- a/processing/src/main/java/org/apache/druid/segment/serde/DictionaryEncodedStringIndexSupplier.java
+++ b/processing/src/main/java/org/apache/druid/segment/serde/DictionaryEncodedStringIndexSupplier.java
@@ -87,7 +87,7 @@ public class DictionaryEncodedStringIndexSupplier implements ColumnIndexSupplier
   public <T> T as(Class<T> clazz)
   {
     if (bitmaps != null) {
-      if (clazz.equals(NullValueIndex.class) && zeroIsNull) {
+      if (clazz.equals(NullValueIndex.class)) {
         final BitmapColumnIndex nullIndex;
         if (zeroIsNull) {
           nullIndex = new SimpleImmutableBitmapIndex(bitmaps.get(0));