… benchmarks/src/test/java/org/apache/druid/benchmark/query/SqlExpressionBenchmark.java
… processing/src/main/java/org/apache/druid/segment/serde/NullValueIndexSupplier.java
… processing/src/main/java/org/apache/druid/segment/virtual/ListFilteredVirtualColumn.java
… processing/src/test/java/org/apache/druid/segment/filter/BoundFilterTest.java
… processing/src/test/java/org/apache/druid/segment/filter/LikeFilterTest.java
… processing/src/test/java/org/apache/druid/segment/filter/SelectorFilterTest.java
… processing/src/test/java/org/apache/druid/segment/virtual/ListFilteredVirtualColumnSelectorTest.java